requirements.txt
flask
pinecone-client
pandas
numpy
tqdm
transformers
torch